Fix wrong type in = operator

+++ include/types/uint48.hpp
@@ -53,7 +53,7 @@ class uint48 {
     uint48(const std::int64_t& a) :
       low(a & 0xFFFFFFFFL), high((a >> 32) & 0xFFFF) {}
 
-    inline uint40& operator=(const uint40& x) {
+    inline uint48& operator=(const uint48& x) {
       if (this != &x) {
         low = x.low;
         high = x.high;
